+++ cincoconfig/support.py
@@ -317,3 +317,38 @@ def asdict(config: Config, virtual: bool = False) -> dict:
         data.update({key: field.__getval__(config) for key, field in fields})
 
     return data
+
+
+def is_default_value(config: Config, key: str) -> bool:
+    '''
+    Check if the given field is set to the default value. This is true when the config value has
+    not been set by a loaded configuration or via the API.
+
+    :param config: configuration object
+    :param key: field key
+    :returns: the field is using the default value
+    '''
+    path, _, key = key.rpartition('.')
+    if path:
+        config = config[path]
+
+    return key in config._default_value_keys
+
+
+def reset_value(config: Config, key: str) -> None:
+    '''
+    Reset a config value back to the default.
+
+    :param config: configuration object
+    :param key: field key
+    '''
+    path, _, key = key.rpartition('.')
+    if path:
+        config = config[path]
+
+    field = config._get_field(key)
+    if not field:
+        raise AttributeError(key)
+
+    field.__setdefault__(config)
+    config._default_value_keys.add(key)